Add specific metric for release recording
To have metric value of `1` for each cluster, accompanied with label for
current release version, add a specific metric for this.

Earlier use of `azure_operator_cluster_status` metric requires
unnecessarily hacky query to function correctly for release statistics
purposes so it's better to have a specific metric for this.